refactor: share drag target detection
